requirements.txt
click==8.1.3
Flask==2.2.2
Flask-WTF==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydub==0.25.1
python-dotenv==0.20.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
